refactor: remove fetch-articles-by-slug on service
